+++ cluster/service/node_test.go
@@ -24,10 +24,10 @@ func testNodeConfig() *Config {
 	}
 }
 
-func testSlaveConfig(idx int) *config.SlaveConfig {
+func testSlaveConfig(idx uint16) *config.SlaveConfig {
 	return &config.SlaveConfig{
 		IP:            "127.0.0.1",
-		Port:          uint64(config.SlavePort + idx),
+		Port:          config.SlavePort + idx,
 		ID:            "S" + string(idx),
 		ChainMaskList: nil,
 	}
